// src/models/AsAdjustmentSearchCondition.ts
export default class AsAdjustmentSearchCondition {
  globalPlantCodeList: string[] = [];
  vehicleProductionPlantCodeList: string[] = [];
  vehicleAssemblyLineCodeList: string[] = [];
  productionPlanMonthList: string[] = [];
  carLineItemCodeList: string[] = [];
  assemblyScheduleAdjustmentInstructionCategoryList: string[] = [];
  assemblyScheduleAdjustmentResultErrorFlagList: string[] = [];
  productionAdjustmentRequestTransferTimeFrom: string = "";
  productionAdjustmentRequestTransferTimeTo: string = "";
  assemblyScheduleAdjustmentUploadUserIdList: string[] = [];
  assemblyScheduleAdjustmentUploadDateTimeFrom: string = "";
  assemblyScheduleAdjustmentUploadDateTimeTo: string = "";
  areaList: string[] = [];
  distributorCodeList: string[] = [];
  vehicleExportDealerCodeList: string[] = [];
  vehicleOrderTypeList: string[] = [];
  sonList: string[] = [];
  baseSpecificationCodeList: string[] = [];
  aicList: string[] = [];
  distributionSpecialOrderCodeList: string[] = [];
  finalLineOffDatePlanFrom: string = "";
  finalLineOffDatePlanTo: string = "";
  transferDestinationAreaList: string[] = [];
  transferDestinationDistributorCodeList: string[] = [];
  transferDestinationVehicleExportDealerCodeList: string[] = [];
  transferDestinationVehicleOrderTypeList: string[] = [];
  transferDestinationSonList: string[] = [];
  transferDestinationBaseSpecificationCodeList: string[] = [];
  transferDestinationAicList: string[] = [];
  transferDestinationDistributionSpecialOrderCodeList: string[] = [];
  transferDestinationFinalLineOffDatePlanFrom: string = "";
  transferDestinationFinalLineOffDatePlanTo: string = "";
}